+++ how-to/workspace-platform-starter/client/src/framework/shapes/app-shapes.ts
@@ -88,11 +88,11 @@ export type PlatformApp = App & {
 };
 
 /**
- *
+ * Are there any preferences you would like to apply when launching this application?
  */
 export interface LaunchPreference {
 	/**
-	 *
+	 * Do you wish to specify a custom height/width that should be used when this application is launched?
 	 */
 	bounds?: {
 		width: number;
@@ -105,7 +105,7 @@ export interface LaunchPreference {
 	defaultCentered?: boolean | undefined;
 
 	/**
-	 *
+	 * Are there any app type specific options you would like to apply?
 	 */
 	options?: {
 		view?: ViewLaunchPreference;
@@ -119,7 +119,7 @@ export interface ViewLaunchPreference {
 	/**
 	 * If specified it indicates you do not want to use a browser window for this view but a platform window.
 	 */
-	hostUrl: string;
+	hostUrl?: string;
 }
 
 /**